# api/swagger/models/api_node_instance_service.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/swag"
)

// APINodeInstanceService api node instance service
// swagger:model apiNodeInstanceService
type APINodeInstanceService struct {

	// address
	Address string `json:"address,omitempty"`

	// engine
	Engine string `json:"engine,omitempty"`

	// engine version
	EngineVersion string `json:"engine_version,omitempty"`

	// id
	ID int64 `json:"id,omitempty"`

	// port
	Port int64 `json:"port,omitempty"`

	// region
	Region string `json:"region,omitempty"`

	// type
	Type string `json:"type,omitempty"`
}

// Validate validates this api node instance service
func (m *APINodeInstanceService) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *APINodeInstanceService) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *APINodeInstanceService) UnmarshalBinary(b []byte) error {
	var res APINodeInstanceService
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}