# api/swagger/models/api_node_remove_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

// APINodeRemoveResponse api node remove response
// swagger:model apiNodeRemoveResponse
type APINodeRemoveResponse interface{}