# api/swagger/models/api_node_remove_service_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

// APINodeRemoveServiceResponse api node remove service response
// swagger:model apiNodeRemoveServiceResponse
type APINodeRemoveServiceResponse interface{}