# api/swagger/node.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "node.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v0/nodes": {
      "get": {
        "operationId": "List",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiNodeListResponse"
            }
          }
        },
        "tags": [
          "Node"
        ]
      }
    },
    "/v0/nodes/{name}": {
      "delete": {
        "operationId": "Remove",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiNodeRemoveResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Node"
        ]
      }
    },
    "/v0/nodes/{name}/services": {
      "delete": {
        "operationId": "RemoveService",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiNodeRemoveServiceResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiNodeRemoveServiceRequest"
            }
          }
        ],
        "tags": [
          "Node"
        ]
      }
    }
  },
  "definitions": {
    "apiNodeInstance": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiNodeInstanceService"
          }
        }
      }
    },
    "apiNodeInstanceService": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "type": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "port": {
          "type": "integer",
          "format": "int64"
        },
        "engine": {
          "type": "string"
        },
        "engine_version": {
          "type": "string"
        }
      }
    },
    "apiNodeListResponse": {
      "type": "object",
      "properties": {
        "instances": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiNodeInstance"
          }
        }
      }
    },
    "apiNodeRemoveResponse": {
      "type": "object"
    },
    "apiNodeRemoveServiceRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "apiNodeRemoveServiceResponse": {
      "type": "object"
    }
  }
}